constraints on share page

+++ ARNotebookCode/ARNotebook/ViewController.swift
@@ -590,7 +590,6 @@ class ViewController:  UIViewController, ARSCNViewDelegate, UIImagePickerControl
      */
     
     func passText(text: String, i: Int = 0) {
-       print("presenting view controller", self.presentingViewController)
         if bookNode != nil && currentPageNode != nil{
             if template == "single"{
                 if selectedTemplate != nil{
@@ -635,7 +634,6 @@ class ViewController:  UIViewController, ARSCNViewDelegate, UIImagePickerControl
                                     node.position = SCNVector3(0,0, 0.01)
                                     node.name = "content"
                                     self.lastNode.append(node)
-                                    print("current page in pass", self.currentPageNode?.name)
                                     let page = self.pages[i-1]
                                     page.childNode(withName: "Top node", recursively: false)?.addChildNode(node)                                   //self.topTempNode?.addChildNode(node)
                                 }).resume()
@@ -657,7 +655,6 @@ class ViewController:  UIViewController, ARSCNViewDelegate, UIImagePickerControl
                                     node.position = SCNVector3(0,0, 0.001)
                                     self.lastNode.append(node)
                                     node.name = "content"
-                                    print("current page in pass:", self.currentPageNode?.name)
                                     let page = self.pages[i-1]
                                     page.childNode(withName: "Bottom node", recursively: false)?.addChildNode(node)
                                     //self.bottomTempNode?.addChildNode(node)
@@ -775,7 +772,12 @@ class ViewController:  UIViewController, ARSCNViewDelegate, UIImagePickerControl
      */
     
     @IBAction func deletePage(_ sender: Any) {
-        self.deletePage()
+        if(accessToWrite){
+            self.deletePage()
+        }
+        else{
+             alert.alert(fromController: self, title:"No Write Access", message:"You are viewing a shared notebook that you do not have write access to. Please continue to use this notebook as read only.")
+        }
     }
     
     func deletePage(){
@@ -1032,10 +1034,6 @@ class ViewController:  UIViewController, ARSCNViewDelegate, UIImagePickerControl
     func attachEventListeners(){
         let postRef = self.ref.child("notebooks/\(self.notebookID)")
         postRef.observe(.childChanged, with: { (snapshot) -> Void in
-            print("something changed")
-            print("snapshot key:", snapshot.key)
-            print("snapshot value:", snapshot.value)
-            print("snapshot children count", snapshot.childrenCount)
             if snapshot.childrenCount == 2 {
                 self.handleSingleChildChange(snapshot: snapshot)
             }
